first 98 bibonacci sequence

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,13 +6,13 @@
 int main(void)
 {
 	int i;
-	int t1 = 1, t2 = 2;
-	int nextTerm = t1 + t2;
+	long t1 = 1, t2 = 2;
+	long nextTerm = t1 + t2;
 
-	printf(" %d, %d, ", t1, t2);
+	printf(" %ld, %ld, ", t1, t2);
 	for (i = 3; i <= 98; ++i)
 	{
-		printf("%d, ", nextTerm);
+		printf("%ld, ", nextTerm);
 		t1 = t2;
 		t2 = nextTerm;
 		nextTerm = t1 + t2;
